rename cookie.rs -> cookie_ext.rs
this is required to re-export the cookie crate
