+++ Cargo.toml
@@ -23,6 +23,7 @@ homepage = "https://FifthTry.com"
 [workspace.dependencies]
 http = "1"
 bytes = "1"
+cookie = { version = "0.18", features = ["percent-encode"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = "1"
