+++ ft-sdk/Cargo.toml
@@ -28,3 +28,7 @@ bytes.workspace = true
 serde_urlencoded.workspace = true
 diesel = { workspace = true, optional = true }
 include_dir.workspace = true
+cookie.workspace = true
+
+[dev-dependencies]
+pretty_assertions = "1"
